refactor: components/board 경로 수정
