+++ src/app/villages/[villageId]/edit/page.jsx
@@ -7,7 +7,7 @@ import {
     villageUpdate,
 } from '../../../../services/villageService';
 import style from './page.module.css';
-import QuillEditor from '../../../../components/common/board/QuillEditor';
+import QuillEditor from '../../../../components/board/QuillEditor';
 
 export default function EditVillagePage() {
     const [data, setData] = useState({
