+++ src/components/board/PostInfoPanel.jsx
@@ -2,11 +2,11 @@
 
 import { useState } from 'react';
 import { useRouter, useParams } from 'next/navigation';
-import Modal from '../Modal';
-import { formatTimeAgo } from '../../../utils/timeFormat';
+import Modal from '../common/Modal';
+import { formatTimeAgo } from '../../utils/timeFormat';
 import style from './PostInfoPanel.module.css';
-import { useAuthStore } from '../../../stores/authStore';
-import AuthorChatButton from '../../chat/AuthorChatButton';
+import { useAuthStore } from '../../stores/authStore';
+import AuthorChatButton from '../chat/AuthorChatButton';
 
 export default function PostInfoPanel({
     data,
