+++ src/components/board/PostInfoPanel.module.css
